JDK-8263322 Added the tests for both cases, when JavaFX was initialized with Application.launch and Platform.startup
